pyproject.toml
[tool.poetry]
name = "yeren_bot"
version = "0.1.0"
description = ""
authors = ["yeren"]

[tool.poetry.dependencies]
python = "^3.8"
graia-ariadne = {extras = ["graia", "alconna"], version = "^0.6.10"}
pymysql = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
# 这里以清华源举例，你也可以使用其他源
name = "tuna-tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = false

